Big update on lots of little things, mostly interfacing
Text message update: Allright, fellas, I reworked quite a bit of the code today to work with interfaces. I didn't really do a good job of cleaning up duplicate code, but we can worry about that once all of these are working and our dungeon is generating everything correctly. 

Burt, the healPotion caused minor problems that I had to tweak to get it to play along with the rest of the interfaces (since it was inheriting before) so implemented the HealBehavior so that it could access all of the parameters it required. Other than a few parameter tweaks there though, it shouldn't be majorly different i don't think. right now the only piece that is still interupting is the "healingPotCount" since heal(character, minHeal, maxHeal, maxHitPoints) since they had to be adapted to DungeonCharacter characters instead of Hero Characters. 

Didn't finish up the view and controller, although I do have them semi-written (and they shouldn't be terrible to implement) so that will be my priority for tomorrow.

I am using the simpler and better encapsulated main (I think i got that from the itemRevamp, but if not, then I pulled it from one of you) to then call the GameManager instead of having all the code out at the beginning. 

Right now the game soft crashes immediately after hero selection/asks if we want to play again and I think that is a result of the main being changed since it has things like "generateDungeon" which i assume will also generate the monsters into whichever rooms along with items. SO ultimately, I think the bug is that we  don't have a monster generating in there right now since we are generating differently then before (where it was created immediately after the monster)

I also rewrote some of the attack methods so that they can be used by either monsters and heroes in all places (unless i missed some which is possible) it was pretty 50/50 before on who we had attacking who and with what parameters). so now all attacks should relatively be attack(opponent, attacker)

I tried to integrate as much of your boy's code into mine as possible, however, i did probably miss something and/or wrote it differently so holler if you have any questions or anything. 

Let me know if you know about the generating monster thing. If you already have an implementation for any of the 6 methods that are still outstanding in GameManager:
generateDungeon(), printInto(), spawnPlayer (might end up being redundant with createPlayer, however, if you are thinking of having that be where we call reading in the name, etc then i'm down with that), checkRoom, gameOver, checkWIn. 



There are more minor changes than those that I made (fixing up all of the existing heros and monsters to only implement interfaces and not extend anything, changed our createHeroes into iDungeonCharacter interfaced switch statement, worked out some of the bugs in the Trap (more or less, i just had it implement an AttackBehavior and messed with the getters/setters, etc for it to work with interfaces 

I think I buffed out some of the issues with healing as well (and the interaction between heal and increasing hitPoints) but haven't been able to test that yet. 

And last but not least, i officially moved everything into the Dungeon package. 

I may get to working on some more tonight, but i have another lab i want to get started tonight so this may be it for the evening. Holler if you have any questions or need anything though. Files shouold be uploading to github as we speak (i'll put this text in the comments too just in case so you can see the changes.